Fix import error, Move error response tests to new file
